Split FilteredIncidentProvider and utility functions to seperate file
