+++ src/components/incidenttable/RealtimeIncidentTable.tsx
@@ -1,77 +1,20 @@
-import React, { useEffect, useState, useContext, useMemo } from "react";
+import React, { useEffect, useState } from "react";
 
 // Api
-import { Incident, IncidentTag } from "../../api";
+import { Incident } from "../../api";
 import { RealtimeService, RealtimeServiceState, RealtimeServiceConfig } from "../../services/RealtimeService";
 
 // Contexts/Hooks
 import { useAlerts } from "../../components/alertsnackbar";
 import { useIncidents } from "../../api/actions";
-import { useIncidentsContext, IncidentsContext } from "../../components/incidentsprovider";
+import { useIncidentsContext } from "../../components/incidentsprovider";
 import { useSelectedFilter } from "../../components/filterprovider";
 
-// Utils
-import { groupBy } from "../../utils";
+// Providers
+import FilteredIncidentsProvider, { matchesFilter } from "../../components/filteredincidentprovider";
 
 // Components
 import { MinimalIncidentTable } from "./IncidentTable";
-import { Tag } from "../../components/tagselector";
-
-// for all different tags "keys", THERE HAS TO BE ONE tag with
-// matching value in incident.tags
-const matchesOnTags = (incident: Incident, tags: Tag[]): boolean => {
-  if (tags.length === 0) {
-    return true;
-  }
-
-  const incidentTagsSet = new Set(incident.tags.map((it: IncidentTag) => it.tag));
-  const tagsGroupedByKey: Map<string, Set<Tag>> = groupBy(tags, (tag: Tag) => tag.key);
-  const returnVal = [...tagsGroupedByKey.keys()].every((tagKey: string) => {
-    const tagsOnKey = tagsGroupedByKey.get(tagKey) || new Set<Tag>();
-    return [...tagsOnKey.values()].some((tag: Tag) => incidentTagsSet.has(tag.original));
-  });
-  return returnVal;
-};
-
-const matchesOnSources = (incident: Incident, sources: number[] | undefined): boolean => {
-  if (sources === undefined || sources.length === 0) {
-    return true;
-  }
-  return sources.some((source: number) => incident.source.pk === source);
-};
-
-const matchesShow = (incident: Incident, show: "open" | "closed" | "both"): boolean => {
-  if (show === "both") return true;
-  const openState = show == "open";
-  return incident.open === openState;
-  // return incident.open ? show === "open" : show === "closed";
-};
-
-const matchesAcked = (incident: Incident, showAcked: boolean): boolean => {
-  if (showAcked) return true;
-  return !incident.acked;
-};
-
-export const FilteredIncidentsProvider = ({
-  filterMatcher,
-  children,
-}: {
-  filterMatcher: ((incident: Incident) => boolean) | undefined;
-  children?: React.ReactNode;
-}) => {
-  const { state, dispatch } = useContext(IncidentsContext);
-
-  const filteredIncidents = useMemo(() => {
-    if (filterMatcher === undefined) return state.incidents;
-    return state.incidents.filter(filterMatcher);
-  }, [filterMatcher, state.incidents]);
-
-  return (
-    <IncidentsContext.Provider value={{ state: { ...state, incidents: filteredIncidents }, dispatch }}>
-      {children}
-    </IncidentsContext.Provider>
-  );
-};
 
 type RealtimeIncidentTablePropsType = {};
 
@@ -145,30 +88,7 @@ const RealtimeIncidentTable = () => {
     // In order for the websockets callbacks to call the updated filter matching
     // function we need to create a new such function every time the filter changes.
     const incidentMatchesFilter = (incident: Incident): boolean => {
-      const filter = { showAcked, show, tags, sourcesById, autoUpdate };
-
-      /*
-      // Useful for debugging
-      const matches = [
-        ["open", incident.open, filter.show, matchesShow(incident, filter.show)],
-        ["acked", incident.acked, filter.showAcked, matchesAcked(incident, filter.showAcked)],
-        ["tags", incident.tags, filter.tags, matchesOnTags(incident, filter.tags)],
-        ["source", incident.source, filter.sourcesById, matchesOnSources(incident, filter.sourcesById)],
-      ];
-
-      console.debug("Checking match on filter: ");
-      matches.forEach(([name, incidentValue, filterValue, match]) => {
-        console.debug(match ? "match" : "fail", `${name}: `, incidentValue, filterValue);
-      });
-
-      return matches.every(([name, incidentValue, filterValue, match]) => match);
-       */
-      return (
-        matchesShow(incident, filter.show) &&
-        matchesAcked(incident, filter.showAcked) &&
-        matchesOnTags(incident, filter.tags) &&
-        matchesOnSources(incident, filter.sourcesById)
-      );
+      return matchesFilter(incident, { showAcked, show, tags, sourcesById });
     };
     setFilterMatcher(() => incidentMatchesFilter);
   }, [showAcked, show, tags, sourcesById, autoUpdate, loadIncidentsFiltered]);
